fix: complete contracts sync
