+++ src/ingestion/ledgers.rs
@@ -1,12 +1,70 @@
-use log::info;
-use prometheus::IntCounter;
-use quasar_entities::ledger;
-use sea_orm::{
-    ActiveModelTrait, ColumnTrait, DatabaseConnection, EntityTrait, QueryFilter, QueryOrder,
-};
-use stellar_node_entities::{ledgerheaders, prelude::Ledgerheaders};
+use super::{sleep, IngestionError};
+use crate::configuration::Ingestion;
+use log::{debug, error, info};
+use prometheus::{IntCounter, Registry};
+use quasar_entities::{ledger, prelude::Ledger};
+use sea_orm::{ActiveModelTrait, ColumnTrait, QueryFilter};
+use sea_orm::{DatabaseConnection, DbErr, EntityTrait, QueryOrder};
+use stellar_node_entities::ledgerheaders;
+use stellar_node_entities::prelude::Ledgerheaders;
 
-use crate::ingestion::IngestionError;
+enum IngestionNeeded {
+    Yes {
+        last_ingested_ledger_sequence: Option<i32>,
+    },
+    No,
+}
+
+async fn new_ledgers_available(
+    node_database: &DatabaseConnection,
+    quasar_database: &DatabaseConnection,
+) -> Result<IngestionNeeded, DbErr> {
+    let last_ingested_ledger_sequence = last_ingested_ledger_sequence(quasar_database).await?;
+    let last_stellar_ledger_sequence = last_stellar_ledger_sequence(node_database).await?;
+
+    let ingestion_needed = if last_ingested_ledger_sequence != last_stellar_ledger_sequence {
+        IngestionNeeded::Yes {
+            last_ingested_ledger_sequence,
+        }
+    } else {
+        IngestionNeeded::No
+    };
+
+    Ok(ingestion_needed)
+}
+
+async fn last_ingested_ledger_sequence(
+    quasar_database: &DatabaseConnection,
+) -> Result<Option<i32>, DbErr> {
+    let last_ingested_ledger = Ledger::find()
+        .order_by_desc(ledger::Column::Sequence)
+        .one(quasar_database)
+        .await?;
+    let last_ingested_ledger_sequence = last_ingested_ledger.map(|ledger| ledger.sequence);
+    Ok(last_ingested_ledger_sequence)
+}
+
+async fn last_stellar_ledger_sequence(
+    node_database: &DatabaseConnection,
+) -> Result<Option<i32>, DbErr> {
+    let last_stellar_ledger = ledgerheaders::Entity::find()
+        .order_by_desc(ledgerheaders::Column::Ledgerseq)
+        .one(node_database)
+        .await?;
+    let last_stellar_ledger_sequence = last_stellar_ledger.and_then(|ledger| ledger.ledgerseq);
+    Ok(last_stellar_ledger_sequence)
+}
+
+fn setup_metrics(
+    metrics: Registry,
+) -> prometheus::core::GenericCounter<prometheus::core::AtomicU64> {
+    let ingested_ledgers_counter =
+        IntCounter::new("ingested_ledgers", "Number of ingested ledgers").unwrap();
+    metrics
+        .register(Box::new(ingested_ledgers_counter.clone()))
+        .expect("Failed to register counter");
+    ingested_ledgers_counter
+}
 
 pub async fn ingest_ledgers(
     node_database: &DatabaseConnection,
@@ -62,4 +120,42 @@ async fn next_ledger_to_ingest(
         .await?;
 
     Ok(next_ledger)
+}
+
+pub async fn ingest(
+    node_database: &DatabaseConnection,
+    quasar_database: &DatabaseConnection,
+    ingestion: &Ingestion,
+    metrics: &Registry,
+) {
+    let ingested_ledgers = setup_metrics(metrics.clone());
+    loop {
+        sleep(ingestion).await;
+
+        // Handle ledgers
+        let ingestion_needed = new_ledgers_available(node_database, quasar_database).await;
+
+        match ingestion_needed {
+            Ok(IngestionNeeded::Yes {
+                last_ingested_ledger_sequence,
+            }) => {
+                debug!("New ledgers available");
+                let ingestion_result = ingest_ledgers(
+                    node_database,
+                    quasar_database,
+                    last_ingested_ledger_sequence,
+                    &ingested_ledgers,
+                )
+                .await;
+
+                if let Err(error) = ingestion_result {
+                    error!("Error while ingesting ledgers: {:?}", error);
+                }
+            }
+            Ok(IngestionNeeded::No) => {}
+            Err(error) => {
+                error!("Error while checking for new ledgers: {}", error);
+            }
+        }
+    }
 }
